% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{exportar_analise_xlsx}
\alias{exportar_analise_xlsx}
\title{Exporta os resultados da análise de agregação para Excel}
\usage{
exportar_analise_xlsx(lista_analise, caminho_arquivo = "resultados_dma.xlsx")
}
\arguments{
\item{lista_analise}{Lista. O objeto retornado por \code{analise_completa_dma()}.}

\item{caminho_arquivo}{String. O nome ou caminho completo do arquivo a ser criado.}
}
\value{
A função não retorna um objeto no R, mas salva o arquivo no disco.
}
\description{
Esta função recebe o objeto de lista gerado pela função
\code{analise_completa_dma()} e salva as tabelas de resumo e
detalhes em abas separadas de um arquivo .xlsx.
}
