% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{plot.agregados_dma}
\alias{plot.agregados_dma}
\title{Plot automatico do objeto agregados_dma}
\usage{
\method{plot}{agregados_dma}(x, amostras_selecionadas = NULL, ...)
}
\description{
Plot automatico do objeto agregados_dma
}
