% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{analise_completa_dma}
\alias{analise_completa_dma}
\title{Executa análise completa e detalhada de agregação}
\usage{
analise_completa_dma(
  dados,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm",
  col_massa = "massa_g"
)
}
\description{
Executa análise completa e detalhada de agregação
}
