% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{calc_dma}
\alias{calc_dma}
\title{Calcula o Diâmetro Médio Ajustado (DMA)}
\usage{
calc_dma(
  df_processado,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm"
)
}
\description{
Calcula o Diâmetro Médio Ajustado (DMA)
}
