% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{plot_diagnostico_amostra}
\alias{plot_diagnostico_amostra}
\title{Gráfico de diagnóstico comparativo para uma amostra}
\usage{
plot_diagnostico_amostra(
  dados,
  amostra_id_alvo,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm",
  col_massa = "massa_g"
)
}
\description{
Gráfico de diagnóstico comparativo para uma amostra
}
