% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{plot_dma}
\alias{plot_dma}
\title{Plota as curvas de retenção e o ajuste do DMA}
\usage{
plot_dma(
  df_processado,
  df_dma,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm"
)
}
\description{
Plota as curvas de retenção e o ajuste do DMA
}
