% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{prep_agregados}
\alias{prep_agregados}
\title{Prepara dados de peneiramento e calcula DMP e DMG}
\usage{
prep_agregados(
  df,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm",
  col_massa = "massa_g"
)
}
\arguments{
\item{df}{Data frame com os dados brutos.}

\item{col_amostra}{String. Nome da coluna de identificacao da amostra.}

\item{col_diametro}{String. Nome da coluna de diametro (mm).}

\item{col_massa}{String. Nome da coluna com a massa (g).}
}
\value{
Uma lista contendo os dados processados e os indices tradicionais.
}
\description{
Prepara dados de peneiramento e calcula DMP e DMG
}
