% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregados.R
\name{prep_agregados}
\alias{prep_agregados}
\title{Prepara dados de peneiramento e calcula DMP e DMG}
\usage{
prep_agregados(
  df,
  col_amostra = "amostra_id",
  col_diametro = "diametro_mm",
  col_massa = "massa_g"
)
}
\description{
Prepara dados de peneiramento e calcula DMP e DMG
}
